[nnc] Move device type checks from isSupported to typesAreSupported
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55025

Something needs to be fixed about the names of these functions,
because they are confusing.

The profiling infrastructure calls `isSupported` to see if it should insert
profiling nodes.

The fuser calls `isSupported` but also `typesAreSupported` to determine if it
can actually fuse the node.

At profiling time, we don't know device types yet, so we can't use device type
checks in `isSupported` or else we'll never profile the node.  So we want to
move those checks into `typesAreSupported`, where we actually have profiling
info available.
ghstack-source-id: 125378651

Differential Revision: [D27454968](https://our.internmc.facebook.com/intern/diff/D27454968/)