[NNAPI] Add unit test
Summary:
This mostly goes op-by-op.  We construct a simple model containing the
op (in various configurations for complex ops) and verify that it can be
converted to NNAPI.  Additionally, if libneuralnetworks is available, we
also run both the eager model and NNAPI model and ensure that their
outputs are equal (allowing for some slight numerical differences).

serializer.py has 94% coverage.  And most of the uncovered lines are
error cases, defensive code, or dead code that I might want to use
later.  prepare.py has 56% coverage, but probably closer to 75-80% if we
could collect coverage from TorchScript.

Test Plan:
Ran tests with NNAPI available.  Made various tweaks to the codebase to
make sure tests properly detected bugs.

ghstack-source-id: f9641c650d797fdae640c988ac8814ea29d04027
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47521